requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai==0.5.0
letta==0.3.0
langgraph==0.0.55
langchain==0.1.0
pandas==2.1.4
numpy==1.26.3
scipy==1.12.0
pymupdf==1.23.8
camelot-py==0.11.0
python-dotenv==1.0.0
duckdb==0.9.2
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
rapidfuzz==3.6.1
pydantic==2.5.3
python-telegram-bot==20.7
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
